==> 3. Introduction to Data Analysis/Lesson 5 Data Analysis Process Case Study 2 - Fule/4 - Assessing Data.ipynb <==
import numpy as np
import pandas as pd

df_fule08 = pd.read_csv("all_alpha_08.csv")

df_fule18 = pd.read_csv("all_alpha_18.csv")


# WORKING WITH THE 2008 DATA

df_fule08

df_fule08.describe()

df_fule08.info()

df_fule08.duplicated().sum()

df_fule08.sample()

# COUNT THE TOTAL ROW WITH ANY MISISNG DATA IN EACH
df_fule08.isnull().any(axis=1).sum()

df_fule08.count()

# WORKING WITH THE 2018 DATA

df_fule18

df_fule18.describe()

df_fule18.info()

df_fule18.isnull().any(axis=1).sum()

df_fule18.duplicated().sum()

df_fule08

f_string ="Ethanol/Gas"
f_data   ='Fuel'
df_loc_in = df_fule08
#df_loc_in = df_fule18
df_loc_in.loc[ df_loc_in[f_data] == f_string ]
