==> 3. Introduction to Data Analysis/Lesson 5 Data Analysis Process Case Study 2 - Fule/6 - Cleaning Column Labels.ipynb <==
# Cleaning Column Labels
   * ### 1. Drop extraneous columns
Drop features that aren't consistent (not present in both datasets) or aren't relevant to our questions. Use pandas' drop function.

## Columns to Drop:
    From 2008 dataset: 'Stnd', 'Underhood ID', 'FE Calc Appr', 'Unadj Cmb MPG'
    From 2018 dataset: 'Stnd', 'Stnd Description', 'Underhood ID', 'Comb CO2'
 * ### 2. Rename Columns
Change the "Sales Area" column label in the 2008 dataset to "Cert Region" for consistency.
Rename all column labels to replace spaces with underscores and convert everything to lowercase. 
(Underscores can be much easier to work with in Python than spaces. 
For example, having spaces wouldn't allow you to use df.column_name instead of df['column_name'] to select columns or use query().
Being consistent with lowercase and underscores also helps make column names easy to remember.)

# Cleaning Column Labels
Use `all_alpha_08.csv` and `all_alpha_18.csv`

import pandas as pd
# load datasets

df_08 = pd.read_csv("all_alpha_08.csv")
df_18 = pd.read_csv("all_alpha_18.csv")

# view 2008 dataset
df_08.head(1)

# view 2018 dataset
df_18.head(1)

### Drop Extraneous Columns

# drop columns from 2008 dataset
df_08.drop(['Stnd', 'Underhood ID', 'FE Calc Appr', 'Unadj Cmb MPG'], axis=1, inplace=True)

# confirm changes
df_08.head(1)

# drop columns from 2018 dataset
df_18.drop(["Stnd", 'Stnd Description', 'Underhood ID', 'Comb CO2'], axis=1 ,inplace=True)



# confirm changes
df_18.head(1)

### Rename Columns

# rename Sales Area to Cert Region
""" df_08["Cert Region"]= df_08["Sales Area"]
df_08.drop(["Sales Area"],axis=1,inplace=True) """
df_08.rename(columns = {"Sales Area" : "Cert Region"}, inplace = True)


# confirm changes
df_08.head(1)

# replace spaces with underscores and lowercase labels for 2008 dataset
df_08.rename(columns=lambda x: x.strip().lower().replace(" ", "_"), inplace=True)

# confirm changes
df_08.head(1)

# replace spaces with underscores and lowercase labels for 2018 dataset
df_18.rename(columns=lambda x:x.strip().lower().replace(" ","_"), inplace=True)

# confirm changes
df_18.head(1)

# confirm column labels for 2008 and 2018 datasets are identical
df_08.columns == df_18.columns

# make sure they're all identical like this
(df_08.columns == df_18.columns).all()

# save new datasets for next section
df_08.to_csv('data_08_v1.csv', index=False)
df_18.to_csv('data_18_v1.csv', index=False)
