==> 3. Introduction to Data Analysis/Lesson 5 Data Analysis Process Case Study 2 - Fule/7 - Filter, Drop Nulls, Dedupe.ipynb <==
# Filter, Drop Nulls, Dedupe
* ### 1. Filter
For consistency, only compare cars certified by California standards.
Filter both datasets using query to select only rows where cert_region is CA. 
Then, drop the cert_region columns, since it will no longer provide any useful information (we'll know every value is 'CA').

 * ### 2. Drop Nulls
Drop any rows in both datasets that contain missing values.

 * ### 3. Dedupe
Drop any duplicate rows in both datasets.



# Filter, Drop Nulls, Dedupe
Use `data_08_v1.csv` and `data_18_v1.csv`. You should've created these data files in the previous section: *Cleaning Column Labels*.

import pandas as pd

# load datasets

df_08 = pd.read_csv("data_08_v1.csv")
df_18 = pd.read_csv("data_18_v1.csv")

# view dimensions of dataset
df_08.shape

# view dimensions of dataset
df_18.shape

## Filter by Certification Region

df_08.head(1)

df_18.head(1)

# filter datasets for rows following California standards
df_08 = df_08[df_08["cert_region"]=="CA"]
df_18 = df_18[df_18["cert_region"]=="CA"]

df_08

df_18

# confirm only certification region is California
df_08['cert_region'].unique()

# confirm only certification region is California
df_18['cert_region'].unique()

# Drop certification region columns form both datasets


df_08.drop(columns=["cert_region"], axis=1 , inplace=True)


df_08.head(1)

df_18.drop(columns=["cert_region"] , axis= 1 , inplace= True)

df_18.head(1)

df_08.shape

df_18.shape

## Drop Rows with Missing Values

# view missing value count for each feature in 2008
df_08.isnull().sum()

# view missing value count for each feature in 2018
df_18.isnull().sum()

# Drop rows with any null values in both datasets



df_08.dropna(inplace=True)
df_18.dropna(inplace=True)

df_08.head(2)

df_18.head(2)

# checks if any of columns in 2008 have null values - should print False
df_08.isnull().sum().any()

# checks if any of columns in 2018 have null values - should print False
df_18.isnull().sum().any()

## Dedupe Data

# print number of duplicates in 2008 and 2018 datasets
df_08.duplicated().sum()


df_18.duplicated().sum()


# drop duplicates in both datasets
df_08.drop_duplicates(inplace=True)
df_18.drop_duplicates(inplace=True)


# print number of duplicates again to confirm dedupe - should both be 0

df_08.duplicated().sum()


df_18.duplicated().sum()


# save progress for the next section
df_08.to_csv('data_08_v2.csv', index=False)
df_18.to_csv('data_18_v2.csv', index=False)
