==> 3. Introduction to Data Analysis/Lesson 5 Data Analysis Process Case Study 2 - Fule/8 - Inspecting Data Types.ipynb <==
# Inspecting Data Types
Use this Jupyter Notebook to inspect the datatypes of features in each dataset and think about what changes should be made to make them practical and consistent (in both datasets). Use this to help you answer the quiz questions below.

# Inspecting Data Types
Use the space below to explore `data_08_v2.csv` and `data_18_v2.csv` to answer the quiz questions below regarding datatypes. You should've created these data files in the previous section: *Filter, Drop Nulls, Dedupe*.

import pandas as pd

df_08 = pd.read_csv("data_08_v2.csv")
df_18 = pd.read_csv("data_18_v2.csv")

df_08.head(2)

df_08.info()

df_18.head(2)

df_18.info()
