==> 3. Introduction to Data Analysis/Lesson 5 Data Analysis Process Case Study 2 - Fule/11 - Fixing Data Types Pt 3.ipynb <==
# Fixing Data Types Part 3
## In this last section, you'll fix datatypes of columns for mpg and greenhouse gas score.

After you complete these final fixes, check the datatypes of all features in both datasets to confirm success for all the changes we specified earlier. Here they are again for your reference:

* ### Fix cyl datatype
    2008: extract int from string.
    2018: convert float to int.

* ### Fix air_pollution_score datatype
    2008: convert string to float.
    2018: convert int to float.

* ### Fix city_mpg, hwy_mpg, cmb_mpg datatypes
    2008 and 2018: convert string to float.

* ### Fix greenhouse_gas_score datatype
    2008: convert from float to int.

## Fix `city_mpg`, `hwy_mpg`, `cmb_mpg` datatypes
    2008 and 2018: convert string to float

Load datasets `data_08_v4.csv` and `data_18_v4.csv`. You should've created these data files in the previous section: *Fixing Data Types Pt 2*.

import pandas as pd

# load datasets

df_08 = pd.read_csv("data_08_v4.csv")
df_18 = pd.read_csv("data_18_v4.csv")

df_08.head(2)

# convert mpg columns to floats
mpg_columns = ["city_mpg" , 'hwy_mpg' , "cmb_mpg"]
for c in mpg_columns:
    df_18[c] = df_18[c].astype(float)
    df_08[c] = df_08[c].astype(float)

## Fix `greenhouse_gas_score` datatype
    2008: convert from float to int

# convert from float to int
df_08['greenhouse_gas_score'] = df_08["greenhouse_gas_score"].astype("int64")

## All the dataypes are now fixed! Take one last check to confirm all the changes.

df_08.dtypes

df_18.dtypes

df_08.dtypes == df_18.dtypes

# Save your final CLEAN datasets as new files!
df_08.to_csv('clean_08.csv', index=False)
df_18.to_csv('clean_18.csv', index=False)
