==> 3. Introduction to Data Analysis/Lesson 5 Data Analysis Process Case Study 2 - Fule/13 - Conclusions & Visuals.ipynb <==
# Conclusions & Visuals
## Draw conclusions and create visuals to communicate results in the Jupyter notebook below! Make sure to address the following questions.

#### Q1: Are more unique models using alternative fuels in 2018 compared to 2008? By how much?
#### Q2: How much have vehicle classes improved in fuel economy (increased in mpg)?
#### Q3: What are the characteristics of SmartWay vehicles? Have they changed over time? (mpg, greenhouse gas)
#### Q4: What features are associated with better fuel economy (mpg)?

# Drawing Conclusions
Use the space below to address questions on datasets `clean_08.csv` and `clean_18.csv`. You should've created these data files in the previous section: *Fixing Data Types Pt 3*.

import pandas as pd

# load datasets

df_08 = pd.read_csv("clean_08.csv")
df_18 = pd.read_csv("clean_18.csv")

### Q1: Are more unique models using alternative sources of fuel? By how much?

df_08.head(1)

df_08.fuel.value_counts()

df_18.fuel.value_counts()

### Q2: How much have vehicle classes improved in fuel economy?  

alt_08 = df_08.query('fuel in ["CNG", "ethanol"]').model.nunique()
alt_08

atl_18 = df_18.query('fuel in ["Electricity","Ethanol"]').model.nunique()
atl_18

### Q3: What are the characteristics of SmartWay vehicles? Have they changed over time?

### Q4: What features are associated with better fuel economy?
