==> 3. Introduction to Data Analysis/Lesson 5 Data Analysis Process Case Study 2 - Fule/15 - Merging Datasets.ipynb <==
# Merging Datasets
***
## 1. Rename 2008 columns to distinguish from 2018 columns after the merge
To do this, use pandas' rename() with a lambda function. See example here.

In the lambda function, take the first 10 characters of the column label and and concatenate it with _2008. (Only take the first 10 characters to prevent really long column names.)

The lambda function should look something like this: lambda x: x[:10] + "_2008"

In your rename, don't forget to specify the parameter columns= when you add the lambda function!
***
## 2. Perform inner merge
To answer the last question, we are only interested in how the same model of car has been updated and how the new model's mpg compares to the old model's mpg.

Perform an inner merge with the left on model_2008 and the right on model. See documentation for pandas' merge here.
***

# Merging Datasets
Use pandas Merges to create a combined dataset from `clean_08.csv` and `clean_18.csv`. You should've created these data files in the previous section: *Fixing Data Types Pt 3*.

import pandas as pd
import numpy as np
import matplotlib.pylab as plt
%matplotlib inline

# load datasets

df_08 = pd.read_csv("clean_08.csv")
df_18 = pd.read_csv("clean_18.csv")

df_08.head(1)

df_18.head(1)

df_08.rename(columns=(lambda x: x[:10] + "_2008" ), inplace=True)

df_08.head(1)

### Create combined dataset

# rename 2008 columns



# view to check names
df_08.head()

# merge datasets
df_combined = df_18.append(df_08, sort=False)

# view to check merge
df_combined.head()

Save the combined dataset

df_combined.to_csv('combined_dataset.csv', index=False)
