==> 3. Introduction to Data Analysis/Lesson 5 Data Analysis Process Case Study 2 - Fule/16 - Results with Merged Dataset.ipynb <==
# Results with Merged Dataset
Use the notebook below to answer the final question with the merged dataset.
***
## Q5: For all of the models that were produced in 2008 that are still being produced now, how much has the mpg improved and which vehicle improved the most?

### Here are the steps for answering this question.
***
### 1. Create a new dataframe, model_mpg, that contain the mean combined mpg values in 2008 and 2018 for each unique model
To do this, group by model and find the mean cmb_mpg_2008 and mean cmb_mpg for each.

### 2. Create a new column, mpg_change, with the change in mpg
Subtract the mean mpg in 2008 from that in 2018 to get the change in mpg

### 3. Find the vehicle that improved the most
Find the max mpg change, and then use query or indexing to see what model it is!
***

# Results with Merged Dataset
#### Q5: For all of the models that were produced in 2008 that are still being produced now, how much has the mpg improved and which vehicle improved the most?
Remember to use your new dataset, `combined_dataset.csv`. You should've created this data file in the previous section: *Merging Datasets*.

import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline

# load dataset

df = pd.read_csv("combined_dataset.csv")
df.head(1)

### 1. Create a new dataframe, `model_mpg`, that contain the mean combined mpg values in 2008 and 2018 for each unique model

To do this, group by `model` and find the mean `cmb_mpg_2008` and mean `cmb_mpg` for each.

model_mpg = df.groupby('model').mean()[['cmb_mpg_2008', 'cmb_mpg']]

model_mpg.head()

### 2. Create a new column, `mpg_change`, with the change in mpg
Subtract the mean mpg in 2008 from that in 2018 to get the change in mpg

model_mpg['mpg_change'] =

model_mpg.head()

### 3. Find the vehicle that improved the most
Find the max mpg change, and then use query or indexing to see what model it is!
